+++ src/proxy-ui-api/frontend/src/components/certificate/CertificateHash.vue
