+++ src/proxy-ui-api/frontend/src/components/certificate/CertificateLine.vue
