+++ src/proxy-ui-api/frontend/src/components/service/WarningDialog.vue
