+++ src/proxy-ui-api/frontend/src/views/Clients/Details/CertificateDialog.vue
