+++ src/proxy-ui-api/frontend/src/views/Clients/Details/ClientDetails.vue
@@ -58,7 +58,7 @@ import Vue from 'vue';
 
 import { mapGetters } from 'vuex';
 import { Permissions } from '@/global';
-import CertificateDialog from '@/components/CertificateDialog.vue';
+import CertificateDialog from './CertificateDialog.vue';
 
 export default Vue.extend({
   components: {
