+++ src/proxy-ui-api/frontend/src/views/Clients/InternalServers/CertificateIcon.vue
